refactor(posthog): migrate flat module into vendor package (T-18)
Move integrations/posthog.py into an integrations/posthog/ package split into
config.py (settings + builders), client.py (HTTP transport + bounce-rate
queries), and verifier.py (credential validation), with a thin __init__.py that
re-exports the public API unchanged.

Add tests/integrations/test_vendor_first_layout.py: a T-1 invariant that fails
if any non-infra flat vendor module reappears under integrations/.

posthog is env-configured analytics with no interactive tools (per registry),
so the package is {config, client, verifier} with no tools/.

Test monkeypatch targets repointed to integrations.posthog.client.*; validator
path test updated to integrations/posthog/verifier.py.
