+++ tests/integrations/test_validator_sentry_capture.py
@@ -93,7 +93,7 @@ CASES: tuple[MigrationCase, ...] = (
     ),
     # posthog
     MigrationCase(
-        "integrations/posthog.py",
+        "integrations/posthog/verifier.py",
         "validate_posthog_config",
         "posthog",
         "validate_posthog_config",
